@@ -1,13 +1,8 @@
-import { getSession } from 'next-auth/client';
-import Document from 'components/Document';
+// import { getSession } from 'next-auth/client';
+import Document from '@/components/Document';
 
-export async function getServerSideProps(ctx) {
-  const session = await getSession(ctx);
-  return {
-    props: {
-      session,
-    },
-  };
-}
+const DocumentPage = ({ user }) => {
+  return <Document user={user} />;
+};
 
-export default Document;
+export default DocumentPage;
